fix some errors and done all test_retrievals

+++ tests/KoPrivateGPT/retrieval/test_hyde_retrieval.py
@@ -9,13 +9,21 @@ from KoPrivateGPT.retrieval import HyDERetrieval, BM25Retrieval
 @pytest.fixture
 def hyde_retrieval():
     test_prompt = "Please write a scientific paper passage to answer the question"
-    bm25_path = os.path.join(test_base_retrieval.root_dir, "resources", "bm25", "bm25_db_retrieval.pkl")
+    bm25_path = os.path.join(test_base_retrieval.root_dir, "resources", "bm25", "test_hyde_retrieval.pkl")
+    pickle_path = os.path.join(test_base_retrieval.root_dir, "resources", "pickle", "test_hyde_retrieval.pkl")
     if not os.path.exists(os.path.dirname(bm25_path)):
         os.makedirs(os.path.dirname(bm25_path))
+    if not os.path.exists(os.path.dirname(pickle_path)):
+        os.makedirs(os.path.dirname(pickle_path))
+
+    test_base_retrieval.ready_pickle_db(pickle_path)
     bm25_retrieval = BM25Retrieval(save_path=bm25_path)
     hyde_retrieval = HyDERetrieval(bm25_retrieval, system_prompt=test_prompt)
     yield hyde_retrieval
-    os.remove(bm25_path)
+    if os.path.exists(pickle_path):
+        os.remove(pickle_path)
+    if os.path.exists(bm25_path):
+        os.remove(bm25_path)
 
 
 def test_hyde_retrieval(hyde_retrieval):
@@ -24,5 +32,5 @@ def test_hyde_retrieval(hyde_retrieval):
     retrieved_ids = hyde_retrieval.retrieve_id(query='What is visconde structure?', top_k=top_k,
                                                model_kwargs={'max_tokens': 64})
     test_base_retrieval.validate_ids(retrieved_ids, top_k)
-
-    # TODO : test retrieve method after making DB linker
+    retrieved_passages = hyde_retrieval.retrieve(query='What is visconde structure?', top_k=top_k)
+    test_base_retrieval.validate_passages(retrieved_passages, top_k)
